Use a configuration directory for each netns rather than a file.
Each directory of the form /etc/default/netns-<netns-name>
represents a network namespace that should be configured.

Each file within this directory represents one instance of a raw
interface, a veth pair, or a veth pair attached to a bridge, much
the same as the previous configuration did. The prefix of each file
determines the type of configuration it contains (raw.*, bridge.*,
veth.*) and multiple files of any type can coexist - they all get
applied to the network namespace with the name given by their parent
directory.